requirements.txt
pygame
telethon
twilio
python-dotenv